refactor: move and rename some methods
